# docs/swagger.yaml
definitions:
...
  controllers.nilaiInput:
    properties:
      indeks:
        type: string
      mahasiswa_id:
        type: integer
      mata_kuliah_id:
        type: integer
      skor:
        type: integer
      user_id:
        type: integer
    required:
    - indeks
    - mahasiswa_id
    - mata_kuliah_id
    - skor
    - user_id
    type: object
...
  models.Nilai:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      indeks:
        type: string
      mahasiswaID:
        type: integer
      mataKuliahID:
        type: integer
      skor:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /nilai:
    get:
      description: Get a list of nilai.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Nilai'
            type: array
      summary: Get all nilai.
      tags:
      - nilai
    post:
      description: Creating a new nilai.
      parameters:
      - description: the body to create a new nilai
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.nilaiInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Nilai'
      security:
      - BearerToken: []
      summary: Create New nilai.
      tags:
      - nilai
  /nilai/{id}:
    delete:
      description: Delete a nilai by id.
      parameters:
      - description: nilai id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one nilai.
      tags:
      - nilai
    get:
      description: Get a nilai by id.
      parameters:
      - description: nilai id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Nilai'
      summary: Get nilai.
      tags:
      - nilai
    patch:
      description: Update nilai by id.
      parameters:
      - description: nilai id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an nilai
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.nilaiInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Nilai'
      security:
      - BearerToken: []
      summary: Update nilai.
      tags:
      - nilai
...
swagger: "2.0"